+++ Source/Util/string+util.h
@@ -9,10 +9,11 @@
 #define string_util_h
 
 #include <stdio.h>
+#include <stdint.h>
 
 void freeStr(char* ptr);
 void strcpy_ignore(char* dst, char* src, char* ignoreSet);
 void strlower(char* str);
-uint strcount(char character, char* str);
+uint64_t strcount(char character, char* str);
 
 #endif /* string_util_h */
